Extract CSP config to its own module
